refactor by gemini cli
