+++ content/browser/BUILD.gn
@@ -3488,6 +3488,7 @@ source_set("browser") {
     }
     if (is_cobalt) {
       deps += [
+        "//starboard/android/shared:starboard_bridge",
         "//third_party/jni_zero:cobalt_for_google3_buildflags",
       ]
     }
